Add category WIP

+++ telegram/request/category.go
@@ -2,7 +2,7 @@ package request
 
 type CreateCategoryRequest struct {
 	CategoryName   string `json:"category_name"`
-	CategoryUserID int    `json:"user_id"`
+	CategoryUserID int64  `json:"user_id"`
 }
 
 type CategoriesResponse struct {
